+++ src/clients/payku.transactions.ts
@@ -23,6 +23,7 @@ import {
   toQueryRecord,
   validateCreateTransactionRequest,
   validateListTransactionsParams,
+  type ValidateCreateTransactionOptions,
 } from "../utils/payku.utils";
 
 export default class PaykuTransactions {
@@ -53,8 +54,9 @@ export default class PaykuTransactions {
 
   private async createTransaction(
     params: PaykuCreateTransactionRequest,
+    options?: ValidateCreateTransactionOptions,
   ): Promise<PaykuCreateTransactionResponse> {
-    validateCreateTransactionRequest(params);
+    validateCreateTransactionRequest(params, options);
 
     try {
       return await this.http.request<PaykuCreateTransactionResponse>({